+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 7.2.0
-- Added support for sharing bindings using `symbol-file` config. )See `README.md`
+- Added support for sharing bindings using `symbol-file` config. (See `README.md`
 and examples/shared_bindings).
 
 # 7.1.0